+++ api/app/config.py
@@ -1,34 +1,81 @@
 """Mimer api default configuration"""
-import os
+import ssl
+from typing import List
 
-# Database connection
-# standard URI has the form:
-# mongodb://[username:password@]host1[:port1][,...hostN[:portN]][/[defaultauthdb][?options]]
-# read more: https://docs.mongodb.com/manual/reference/connection-string/
-DATABASE_NAME = os.getenv("DATABASE_NAME", "bonsai")
-DB_HOST = os.getenv("DB_HOST", "mongodb")
-DB_PORT = os.getenv("DB_PORT", "27017")
-MONGODB_URI = f"mongodb://{DB_HOST}:{DB_PORT}/{DATABASE_NAME}"
-MAX_CONNECTIONS = int(os.getenv("MAX_CONNECTIONS", "10"))
-MIN_CONNECTIONS = int(os.getenv("MIN_CONNECTIONS", "10"))
+from pydantic_settings import BaseSettings
 
-# Redis connection
-REDIS_HOST = os.getenv("REDIS_HOST", "redis")
-REDIS_PORT = os.getenv("REDIS_PORT", "6379")
+ssl_defaults = ssl.get_default_verify_paths()
 
-# Reference genome and annotations for IGV
-REFERENCE_GENOMES_DIR = os.getenv("REFERENCE_GENOMES_DIR", "/tmp/reference_genomes")
-ANNOTATIONS_DIR = os.getenv("ANNOTATIONS_DIR", "/tmp/annotations")
 
-# Configure allowed origins (CORS) for development. Origins are a comma seperated list.
-# https://fastapi.tiangolo.com/tutorial/cors/
-allowed_origins = os.getenv("ALLOWED_ORIGINS", "").split(",")
+class Settings(BaseSettings):
+    """API configuration."""
+    # Configure allowed origins (CORS) for development. Origins are a comma seperated list.
+    # https://fastapi.tiangolo.com/tutorial/cors/
+    allowed_origins: List[str] = []
+
+    # Database connection
+    # standard URI has the form:
+    # mongodb://[username:password@]host1[:port1][,...hostN[:portN]][/[defaultauthdb][?options]]
+    # read more: https://docs.mongodb.com/manual/reference/connection-string/
+    database_name: str = "bonsai"
+    db_host: str = "mongodb"
+    db_port: str = "27017"
+    mongodb_uri: str = f"mongodb://{db_host}:{db_port}/{database_name}"
+    max_connections: int = 10
+    min_connections: int = 10
+
+    # Redis connection
+    redis_host: str = "redis"
+    redis_port: str = "6379"
+
+    # Reference genome and annotations for IGV
+    reference_genomes_dir: str = "/tmp/reference_genomes"
+    annotations_dir: str = "/tmp/annotations"
+    # authentication options
+    secret_key: str = "not-so-secret"  # openssl rand -hex 32
+    access_token_expire_minutes: int = 180  # expiration time for accesst token
+    # LDAP login Settings
+    # If LDAP is not configured it will fallback on local authentication
+    ldap_search_attr: str = "mail"
+    ldap_search_filter: str | None = None
+    ldap_base_dn: str | None = None
+    # ldap server
+    ldap_host: str | None = None
+    ldap_port: int = 1389
+    ldap_bind_dn: str | None = None
+    ldap_secret: str | None = None
+    ldap_connection_timeout: int = 10
+    ldap_read_only: bool = False
+    ldap_valid_names: str | None = None
+    ldap_private_key_password: str | None = None
+    ldap_raise_exceptions: bool = False
+    ldap_user_login_attr: str = "mail"
+    force_attribute_value_as_list: bool = False
+    # ldap tls
+    ldap_use_ssl: bool = False
+    ldap_use_tls: bool = True
+    ldap_tls_version: int = ssl.PROTOCOL_TLSv1
+    ldap_require_cert: int = ssl.CERT_REQUIRED
+    ldap_client_private_key: str | None = None
+    ldap_client_cert: str | None = None
+    # ldap ssl
+    ldap_ca_certs_file: str | None = ssl_defaults.cafile
+    ldap_ca_certs_path: str | None = ssl_defaults.capath
+    ldap_ca_certs_data: str | None = None
+
+    @property
+    def use_ldap_auth(self) -> bool:
+        """Return True if LDAP authentication is enabled.
+
+        :return: Return True if LDAP authentication is enabled
+        :rtype: bool
+        """
+        return self.ldap_host is not None
+
 
 # to get a string like this run:
 # openssl rand -hex 32
-SECRET_KEY = "not-so-secret"
 ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 180
 
 # Definition of user roles
 USER_ROLES = {
@@ -53,18 +100,8 @@ USER_ROLES = {
         "locations:write",
     ],
     "uploader": [
-        "groups:write"
-        "samples:write",
+        "groups:write" "samples:write",
     ],
 }
 
-# Configure authentication method used
-# If LDAP is not configured it will fallback on local authentication
-
-# LDAP login Settings
-# LDAP_HOST = "localhost"
-# LDAP_PORT = 389
-# LDAP_BASE_DN = 'cn=admin,dc=example,dc=com
-# LDAP_USER_LOGIN_ATTR = "mail"
-# LDAP_USE_SSL = False
-# LDAP_USE_TLS = True
+settings = Settings()